api_gateway/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
astunparse==1.6.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.110.0
graphql-core==3.2.3
h11==0.14.0
httptools==0.6.1
idna==3.6
pydantic==2.6.4
pydantic-core==2.16.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
strawberry-graphql==0.221.1
typing-extensions==4.10.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
